+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Compatibility:
 * Fix `Enumerator.product` size for empty enumerators (#4249, @earlopain)
 * Consistent coercion error messages (#4250, @earlopain)
 * `IO.select` now accepts `Float::INFINITY` as a timeout argument (#4252, @haukot)
+* Implement `Math.log1p` and `Math.expm1` (#4255, @haukot).
 
 Performance: